+++ src/bsoncxx/include/bsoncxx/v_noabi/bsoncxx/builder/basic/impl.hpp
@@ -34,7 +34,7 @@ detail::requires_t<void, detail::is_invocable<T, sub_document>> generic_append(c
     core->close_document();
 }
 
-template <typename T, typename Placeholder = void> // placeholder 'void' for VS2015 compat
+template <typename T>
 detail::requires_t<void, detail::is_invocable<T, sub_array>> generic_append(core* core, T&& func) {
     core->open_array();
     detail::invoke(std::forward<T>(func), sub_array{core});