+++ src/bsoncxx/include/bsoncxx/v_noabi/bsoncxx/builder/stream/value_context.hpp
@@ -113,7 +113,7 @@ class value_context {
     operator single_context();
 
 #if !defined(_MSC_VER) && !defined(__INTEL_COMPILER)
-    // TODO(MSVC): Causes an ICE under VS2015U1
+    // TODO(MSVC): Causes an ICE under Visual Studio.
     static_assert(
         std::is_same<value_context, decltype(std::declval<value_context>() << 1 << "str")>::value,
         "value_context must be templatized on a key_context");